+++ packages/scrollbar/src/Component.tsx
@@ -87,6 +87,11 @@ export type ScrollbarProps = {
     verticalBarClassName?: string;
 
     /**
+     * Дополнительный класс контейнера горизонтальной полосы прокрутки
+     */
+    horizontalBarClassName?: string;
+
+    /**
      * Обработчик изменения скролла
      */
     onContentScroll?: (e: Event) => void;
@@ -150,6 +155,7 @@ export const Scrollbar = React.forwardRef<HTMLDivElement, ScrollbarProps>(
             widthPropName = 'minWidth',
             maskProps,
             verticalBarClassName,
+            horizontalBarClassName,
             onContentScroll,
             verticalBarRef,
             ...htmlAttributes
@@ -299,7 +305,9 @@ export const Scrollbar = React.forwardRef<HTMLDivElement, ScrollbarProps>(
                     </div>
                     <div className={classNames.placeholder} />
                 </div>
-                <div className={cn(classNames.track, classNames.horizontal)}>
+                <div
+                    className={cn(classNames.track, classNames.horizontal, horizontalBarClassName)}
+                >
                     <div className={classNames.scrollbar} />
                 </div>
                 <div