+++ packages/universal-modal/CHANGELOG.md
@@ -1,5 +1,27 @@
 # Change Log
 
+## 1.1.5
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.18.0
+    -   scrollbar@3.4.0
+    -   base-modal@5.9.4
+    -   button@11.11.10
+    -   navigation-bar-private@0.8.7
+    -   typography@4.14.3
+
+## 1.1.4
+
+### Patch Changes
+
+<sup><time>17.06.2025</time></sup>
+
+### [#1743](https://github.com/core-ds/core-components/pull/1743)
+
+-   Открыт для использования пропс `disableBackdropClick` в десктопной версии
+
 ## 1.1.3
 
 ### Patch Changes